+++ tests/VectorTraits.Tests/Impl/IWVectorTraits512Test/Vector512Tests_YB.cs
@@ -14,7 +14,6 @@ namespace Zyl.VectorTraits.Tests.Impl.IWVectorTraits512Test {
     public class Vector512Tests_YB {
 #if NET8_0_OR_GREATER
 
-
         [TestCase]
         public void YBitToByteTest() {
             IReadOnlyList<IWVectorTraits512> instances = Vector512s.TraitsInstances;
@@ -25,26 +24,26 @@ namespace Zyl.VectorTraits.Tests.Impl.IWVectorTraits512Test {
                     Console.WriteLine($"{instance.GetType().Name}: {instance.GetUnsupportedMessage()}");
                 }
             }
-            var funcList = Vector512s.GetSupportedMethodList<Func<uint, Vector512<byte>>>("YBitToByte_Basic", "YBitToByte_Shuffle", "YBitToByte_Widen");
+            var funcList = Vector512s.GetSupportedMethodList<Func<ulong, Vector512<byte>>>("YBitToByte_Basic", "YBitToByte_Shuffle", "YBitToByte_Widen");
             foreach (var func in funcList) {
                 Console.WriteLine("{0}: OK", ReflectionUtil.GetShortNameWithType(func.Method));
             }
             // samples.
-            List<uint> samples = new List<uint>();
+            List<ulong> samples = new List<ulong>();
             int bitCount = Vector512<byte>.Count;
             int bitClamped = Math.Min(bitCount, 4);
             int iMax = (1 << bitClamped) - 1;
             for (int i = 0; i <= iMax; ++i) {
-                samples.Add((uint)i);
+                samples.Add((ulong)i);
             }
             if (bitClamped < bitCount) {
                 for (int i = bitClamped; i < bitCount; ++i) {
                     iMax <<= 1;
-                    samples.Add((uint)iMax);
+                    samples.Add((ulong)iMax);
                 }
             }
             // run.
-            foreach (uint value in samples) {
+            foreach (ulong value in samples) {
                 Vector512<byte> expected = Vector512s.YBitToByte(value);
                 //Console.WriteLine(VectorTextUtil.Format("YBitToByte({0}):\t{1}", value, expected));
                 foreach (IWVectorTraits512 instance in instances) {
@@ -73,21 +72,21 @@ namespace Zyl.VectorTraits.Tests.Impl.IWVectorTraits512Test {
                 }
             }
             // samples.
-            List<uint> samples = new List<uint>();
+            List<ulong> samples = new List<ulong>();
             int bitCount = Vector512<short>.Count;
             int bitClamped = Math.Min(bitCount, 4);
             int iMax = (1 << bitClamped) - 1;
             for (int i = 0; i <= iMax; ++i) {
-                samples.Add((uint)i);
+                samples.Add((ulong)i);
             }
             if (bitClamped < bitCount) {
                 for (int i = bitClamped; i < bitCount; ++i) {
                     iMax <<= 1;
-                    samples.Add((uint)iMax);
+                    samples.Add((ulong)iMax);
                 }
             }
             // run.
-            foreach (uint value in samples) {
+            foreach (ulong value in samples) {
                 Vector512<short> expected = Vector512s.YBitToInt16(value);
                 //Console.WriteLine(VectorTextUtil.Format("YBitToInt16({0}):\t{1}", value, expected));
                 foreach (IWVectorTraits512 instance in instances) {
@@ -109,21 +108,21 @@ namespace Zyl.VectorTraits.Tests.Impl.IWVectorTraits512Test {
                 }
             }
             // samples.
-            List<uint> samples = new List<uint>();
+            List<ulong> samples = new List<ulong>();
             int bitCount = Vector512<int>.Count;
             int bitClamped = Math.Min(bitCount, 4);
             int iMax = (1 << bitClamped) - 1;
             for (int i = 0; i <= iMax; ++i) {
-                samples.Add((uint)i);
+                samples.Add((ulong)i);
             }
             if (bitClamped < bitCount) {
                 for (int i = bitClamped; i < bitCount; ++i) {
                     iMax <<= 1;
-                    samples.Add((uint)iMax);
+                    samples.Add((ulong)iMax);
                 }
             }
             // run.
-            foreach (uint value in samples) {
+            foreach (ulong value in samples) {
                 Vector512<int> expected = Vector512s.YBitToInt32(value);
                 //Console.WriteLine(VectorTextUtil.Format("YBitToInt32({0}):\t{1}", value, expected));
                 foreach (IWVectorTraits512 instance in instances) {
@@ -145,21 +144,21 @@ namespace Zyl.VectorTraits.Tests.Impl.IWVectorTraits512Test {
                 }
             }
             // samples.
-            List<uint> samples = new List<uint>();
+            List<ulong> samples = new List<ulong>();
             int bitCount = Vector512<long>.Count;
             int bitClamped = Math.Min(bitCount, 4);
             int iMax = (1 << bitClamped) - 1;
             for (int i = 0; i <= iMax; ++i) {
-                samples.Add((uint)i);
+                samples.Add((ulong)i);
             }
             if (bitClamped < bitCount) {
                 for (int i = bitClamped; i < bitCount; ++i) {
                     iMax <<= 1;
-                    samples.Add((uint)iMax);
+                    samples.Add((ulong)iMax);
                 }
             }
             // run.
-            foreach (uint value in samples) {
+            foreach (ulong value in samples) {
                 Vector512<long> expected = Vector512s.YBitToInt64(value);
                 //Console.WriteLine(VectorTextUtil.Format("YBitToInt64({0}):\t{1}", value, expected));
                 foreach (IWVectorTraits512 instance in instances) {
